chrono: merge from upstream
