+++ HLTrigger/Timer/test/omp_get_wtime.h
@@ -1,7 +1,9 @@
+#ifndef __clang__
+// CLANG does not support OpenMP
+
 #ifndef omp_get_wtime_h
 #define omp_get_wtime_h
 
-#ifndef __clang__
 // C++ standard headers
 #include <chrono>
 
@@ -25,6 +27,7 @@ struct clock_omp_get_wtime
     return time_point(duration( seconds ));
   }
 };
-#endif
 
 #endif // omp_get_wtime_h
+
+#endif // ! __clang__
