+++ HLTrigger/Timer/test/x86_tsc_clock.h
@@ -9,10 +9,9 @@
 
 // for rdtscp, rdtscp, lfence, mfence
 #include <x86intrin.h>
-#include "x86_tsc.h"
 
-// for native_duration, tsc_tick, etc.
-#include "native.h"
+// for tsc_tick, etc.
+#include "x86_tsc.h"
 #include "x86_tsc_tick.h"
 
 // TSC-based clock, using rdtsc (non-serialising)
@@ -105,95 +104,30 @@ struct clock_rdtscp
 };
 
 
-// TSC-based clock with native duration, using rdtsc (non-serialising)
-struct clock_rdtsc_native
+#if defined __GLIBC__ && (__GLIBC__ > 2) || (__GLIBC__ == 2 && __GLIBC_MINOR__ >= 11)
+// IFUNC support requires GLIBC >= 2.11.1
+// TSC-based clock, determining at run-time the best strategy to serialise the reads from the TSC
+struct clock_serialising_rdtsc
 {
-  // std::chrono-like native interface
-  typedef native_duration<int64_t, tsc_tick>                            duration;
+  // std::chrono interface
+  typedef std::chrono::nanoseconds                                      duration;
   typedef duration::rep                                                 rep;
   typedef duration::period                                              period;
-  typedef native_time_point<clock_rdtsc_native, duration>               time_point;
+  typedef std::chrono::time_point<clock_rdtscp, duration>               time_point;
 
   static const bool is_steady;
   static const bool is_available;
 
   static time_point now() noexcept
   {
-    rep        ticks = rdtsc();
-    duration   d(ticks);
-    time_point t(d);
-    return t;
+    int64_t    ticks = serialising_rdtsc();
+    rep        ns    = tsc_tick::to_nanoseconds(ticks);
+    time_point time  = time_point(duration(ns));
+    return time;
   }
 };
+#endif // defined __GLIBC__ && (__GLIBC__ > 2) || (__GLIBC__ == 2 && __GLIBC_MINOR__ >= 11)
 
-
-// TSC-based clock with native duration, using lfence as serialising instruction
-struct clock_rdtsc_lfence_native
-{
-  // std::chrono-like native interface
-  typedef native_duration<int64_t, tsc_tick>                            duration;
-  typedef duration::rep                                                 rep;
-  typedef duration::period                                              period;
-  typedef native_time_point<clock_rdtsc_lfence_native, duration>        time_point;
-
-  static const bool is_steady;
-  static const bool is_available;
-
-  static time_point now() noexcept
-  {
-    _mm_lfence();
-    rep        ticks = rdtsc();
-    duration   d(ticks);
-    time_point t(d);
-    return t;
-  }
-};
-
-
-// TSC-based clock with native duration, using mfence as serialising instruction
-struct clock_rdtsc_mfence_native
-{
-  // std::chrono-like native interface
-  typedef native_duration<int64_t, tsc_tick>                            duration;
-  typedef duration::rep                                                 rep;
-  typedef duration::period                                              period;
-  typedef native_time_point<clock_rdtsc_mfence_native, duration>        time_point;
-
-  static const bool is_steady;
-  static const bool is_available;
-
-  static time_point now() noexcept
-  {
-    _mm_mfence();
-    rep        ticks = rdtsc();
-    duration   d(ticks);
-    time_point t(d);
-    return t;
-  }
-};
-
-
-// TSC-based clock with native duration, using rdtscp as serialising instruction
-struct clock_rdtscp_native
-{
-  // std::chrono-like native interface
-  typedef native_duration<int64_t, tsc_tick>                            duration;
-  typedef duration::rep                                                 rep;
-  typedef duration::period                                              period;
-  typedef native_time_point<clock_rdtscp_native, duration>              time_point;
-
-  static const bool is_steady;
-  static const bool is_available;
-
-  static time_point now() noexcept
-  {
-    unsigned int id;
-    rep        ticks = rdtscp(& id);
-    duration   d(ticks);
-    time_point t(d);
-    return t;
-  }
-};
 
 #endif // x86_tsc_clock_h
